backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
boto3
imageio-ffmpeg
requests
moviepy
Pillow